refactor: consolidate helper
