+++ src/views/product-landing/helpers/get-icon-cards.tsx
@@ -3,7 +3,7 @@ import { IconDocs16 } from '@hashicorp/flight-icons/svg-react/docs-16'
 import { IconDownload16 } from '@hashicorp/flight-icons/svg-react/download-16'
 import { IconLearn16 } from '@hashicorp/flight-icons/svg-react/learn-16'
 import { ProductData } from 'types/products'
-import { getIsEnabledProductIntegrations } from 'lib/integrations/get-product-slugs-with-integrations'
+import { getIsEnabledProductIntegrations } from 'lib/integrations/get-is-enabled-product-integrations'
 
 export function getIconCards(product: ProductData) {
 	const iconCards = [
